Messages for remote items
